added "remnant storage"
remnant storage is just my way of saying "stuff stored in db, yoink that stuff out so data on page is the same again" so all drops will appear to have the same colour, radius, and position. New drops will still be added randomly. Allows for the new drops to be more visible. Still haven't added the functionality for them to fade out slowly but that's because I don't think that's possible on the HTML5 canvas with Javascript. Might resort to some library. Got some recommendations?